vectorscope waveform waveform: move some code to common.c
