+++ src/vectorscope.c
@@ -1,8 +1,6 @@
 #include <obs-module.h>
 #include <graphics/image-file.h>
 #include <util/platform.h>
-#include <util/threading.h>
-#include <obs-frontend-api.h>
 #include "plugin-macros.generated.h"
 #include "obs-convenience.h"
 #include "common.h"
@@ -13,10 +11,6 @@
 #define PROFILE_START(x) profile_start(x)
 #define PROFILE_END(x) profile_end(x)
 static const char *prof_render_name = "vss_render";
-static const char *prof_render_name_b = "vss_render_bypass";
-static const char *prof_render_target_name = "render_target";
-static const char *prof_convert_yuv_name = "convert_yuv";
-static const char *prof_stage_surface_name = "stage_surface";
 static const char *prof_draw_vectorscope_name = "draw_vectorscope";
 static const char *prof_draw_name = "draw";
 static const char *prof_draw_graticule_name = "graticule";
@@ -26,13 +20,12 @@ static const char *prof_draw_graticule_name = "graticule";
 #endif // ! ENABLE_PROFILE
 
 #define VS_SIZE 256
-#define SOURCE_CHECK_NS 3000000000
 #define N_GRATICULES 18
 #define GRATICULES_IQ 256
 #define GRATICULES_COLOR_MASK 3
 #define SKIN_TONE_LINE 0x0054FF // BGR
 
-gs_effect_t *vss_effect = NULL;
+extern gs_effect_t *cm_rgb2yuv_effect;
 
 #define RGB2Y_601(r, g, b) ((+306*(r) +601*(g) +117*(b))/1024 +  0)
 #define RGB2U_601(r, g, b) ((-150*(r) -296*(g) +448*(b))/1024 +128)
@@ -44,37 +37,21 @@ gs_effect_t *vss_effect = NULL;
 
 struct vss_source
 {
-	obs_source_t *self;
-	gs_texrender_t *texrender;
-	gs_texrender_t *texrender_uv;
-	gs_stagesurf_t* stagesurface;
-	uint32_t known_width;
-	uint32_t known_height;
+	struct cm_source cm;
 
 	gs_texture_t *tex_vs;
 	uint8_t *tex_buf;
-
-	pthread_mutex_t target_update_mutex;
-	uint64_t target_check_time;
-	obs_weak_source_t *weak_target;
-	char *target_name;
 
 	gs_image_file_t graticule_img;
 	gs_vertbuffer_t *graticule_vbuf;
 	gs_vertbuffer_t *graticule_line_vbuf;
 
-	int target_scale;
 	int intensity;
 	int graticule;
 	int graticule_color;
 	int graticule_skintone_color;
 	int colorspace;
-	int colorspace_calc; // get from ovi if auto
 	bool update_graticule;
-	bool bypass_vectorscope;
-
-	bool rendered;
-	bool enumerating; // not thread safe but I have no other idea.
 };
 
 static const char *vss_get_name(void *unused)
@@ -89,18 +66,8 @@ static void *vss_create(obs_data_t *settings, obs_source_t *source)
 {
 	struct vss_source *src = bzalloc(sizeof(struct vss_source));
 
-	src->self = source;
-	obs_enter_graphics();
-	src->texrender = gs_texrender_create(GS_BGRA, GS_ZS_NONE);
-	src->texrender_uv = gs_texrender_create(GS_BGRA, GS_ZS_NONE);
-	if (!vss_effect) {
-		char *f = obs_module_file("vectorscope.effect");
-		vss_effect = gs_effect_create_from_file(f, NULL);
-		if (!vss_effect)
-			blog(LOG_ERROR, "Cannot load '%s'", f);
-		bfree(f);
-	}
-	obs_leave_graphics();
+	src->cm.flags = CM_FLAG_CONVERT_UV;
+	cm_create(&src->cm, settings, source);
 
 	{
 		// The file is generated by
@@ -115,8 +82,6 @@ static void *vss_create(obs_data_t *settings, obs_source_t *source)
 		bfree(f);
 	}
 
-	pthread_mutex_init(&src->target_update_mutex, NULL);
-
 	vss_update(src, settings);
 
 	return src;
@@ -127,10 +92,6 @@ static void vss_destroy(void *data)
 	struct vss_source *src = data;
 
 	obs_enter_graphics();
-	gs_stagesurface_destroy(src->stagesurface);
-	gs_texrender_destroy(src->texrender);
-	gs_texrender_destroy(src->texrender_uv);
-
 	gs_texture_destroy(src->tex_vs);
 	bfree(src->tex_buf);
 	gs_image_file_free(&src->graticule_img);
@@ -138,37 +99,14 @@ static void vss_destroy(void *data)
 	gs_vertexbuffer_destroy(src->graticule_line_vbuf);
 	obs_leave_graphics();
 
-	bfree(src->target_name);
-	pthread_mutex_destroy(&src->target_update_mutex);
-	obs_weak_source_release(src->weak_target);
-
+	cm_destroy(&src->cm);
 	bfree(src);
 }
 
 static void vss_update(void *data, obs_data_t *settings)
 {
 	struct vss_source *src = data;
-	obs_weak_source_t *weak_source_old = NULL;
-
-	const char *target_name = obs_data_get_string(settings, "target_name");
-	if (!src->target_name || strcmp(target_name, src->target_name)) {
-		pthread_mutex_lock(&src->target_update_mutex);
-		bfree(src->target_name);
-		src->target_name = bstrdup(target_name);
-		weak_source_old = src->weak_target;
-		src->weak_target = NULL;
-		src->target_check_time = os_gettime_ns() - SOURCE_CHECK_NS;
-		pthread_mutex_unlock(&src->target_update_mutex);
-	}
-
-	if (weak_source_old) {
-		obs_weak_source_release(weak_source_old);
-		weak_source_old = NULL;
-	}
-
-	src->target_scale = (int)obs_data_get_int(settings, "target_scale");
-	if (src->target_scale<1)
-		src->target_scale = 1;
+	cm_update(&src->cm, settings);
 
 	src->intensity = (int)obs_data_get_int(settings, "intensity");
 	if (src->intensity<1)
@@ -194,7 +132,6 @@ static void vss_update(void *data, obs_data_t *settings)
 		src->colorspace = colorspace;
 		src->update_graticule = 1;
 	}
-	src->bypass_vectorscope = obs_data_get_bool(settings, "bypass_vectorscope");
 }
 
 static void vss_get_defaults(obs_data_t *settings)
@@ -212,10 +149,8 @@ static obs_properties_t *vss_get_properties(void *data)
 	obs_property_t *prop;
 	props = obs_properties_create();
 
-	prop = obs_properties_add_list(props, "target_name", obs_module_text("Source"), OBS_COMBO_TYPE_LIST, OBS_COMBO_FORMAT_STRING);
-	property_list_add_sources(prop, src ? src->self : NULL);
+	cm_get_properties(&src->cm, props);
 
-	obs_properties_add_int(props, "target_scale", obs_module_text("Scale"), 1, 128, 1);
 	obs_properties_add_int(props, "intensity", obs_module_text("Intensity"), 1, 255, 1);
 	prop = obs_properties_add_list(props, "graticule", obs_module_text("Graticule"), OBS_COMBO_TYPE_LIST, OBS_COMBO_FORMAT_INT);
 	obs_property_list_add_int(prop, "None", 0);
@@ -231,35 +166,19 @@ static obs_properties_t *vss_get_properties(void *data)
 	obs_property_list_add_int(prop, "601", 1);
 	obs_property_list_add_int(prop, "709", 2);
 
-	obs_properties_add_bool(props, "bypass_vectorscope", obs_module_text("Bypass"));
-
 	return props;
 }
 
 static uint32_t vss_get_width(void *data)
 {
 	struct vss_source *src = data;
-	return src->bypass_vectorscope ? src->known_width : VS_SIZE;
+	return src->cm.bypass ? src->cm.known_width : VS_SIZE;
 }
 
 static uint32_t vss_get_height(void *data)
 {
 	struct vss_source *src = data;
-	return src->bypass_vectorscope ? src->known_height : VS_SIZE;
-}
-
-static void vss_enum_sources(void *data, obs_source_enum_proc_t enum_callback, void *param)
-{
-	struct vss_source *src = data;
-	if (src->enumerating)
-		return;
-	src->enumerating = 1;
-	obs_source_t *target = obs_weak_source_get_source(src->weak_target);
-	if (target) {
-		enum_callback(src->self, target, param);
-		obs_source_release(target);
-	}
-	src->enumerating = 0;
+	return src->cm.bypass ? src->cm.known_height : VS_SIZE;
 }
 
 static inline void vss_draw_vectorscope(struct vss_source *src, uint8_t *video_data, uint32_t video_line)
@@ -271,8 +190,8 @@ static inline void vss_draw_vectorscope(struct vss_source *src, uint8_t *video_d
 	for (int i=0; i<VS_SIZE*VS_SIZE; i++)
 		dbuf[i] = 0;
 
-	const uint32_t height = src->known_height;
-	const uint32_t width = src->known_width;
+	const uint32_t height = src->cm.known_height;
+	const uint32_t width = src->cm.known_width;
 	uint8_t *vd = video_data;
 	uint32_t vd_add = video_line - width*4;
 	for (uint32_t y=0; y<height; y++) {
@@ -292,105 +211,6 @@ static inline void vss_draw_vectorscope(struct vss_source *src, uint8_t *video_d
 		src->tex_vs = gs_texture_create(VS_SIZE, VS_SIZE, GS_R8, 1, (const uint8_t**)&src->tex_buf, GS_DYNAMIC);
 	else
 		gs_texture_set_image(src->tex_vs, src->tex_buf, VS_SIZE, false);
-}
-
-static void vss_render_target(struct vss_source *src)
-{
-	if (src->rendered)
-		return;
-	src->rendered = 1;
-
-	obs_source_t *target = src->weak_target ? obs_weak_source_get_source(src->weak_target) : NULL;
-	if (!target && *src->target_name)
-		return;
-
-	int target_width, target_height;
-	if (target) {
-		target_width = obs_source_get_width(target);
-		target_height = obs_source_get_height(target);
-	}
-	else {
-		struct obs_video_info ovi;
-		obs_get_video_info(&ovi);
-		target_width = ovi.base_width;
-		target_height = ovi.base_height;
-	}
-	int width = target_width / src->target_scale;
-	int height = target_height / src->target_scale;
-	if (width<=0 || height<=0)
-		goto end;
-
-	PROFILE_START(prof_render_target_name);
-
-	gs_texrender_reset(src->texrender);
-	if (gs_texrender_begin(src->texrender, width, height)) {
-		struct vec4 background;
-		vec4_zero(&background);
-
-		gs_clear(GS_CLEAR_COLOR, &background, 0.0f, 0);
-		gs_ortho(0.0f, (float)target_width, 0.0f, (float)target_height, -100.0f, 100.0f);
-
-		gs_blend_state_push();
-		if (target) {
-			gs_blend_function(GS_BLEND_ONE, GS_BLEND_ZERO);
-			obs_source_video_render(target);
-		}
-		else
-			obs_render_main_texture();
-
-		gs_texrender_end(src->texrender);
-
-		if (width != src->known_width || height != src->known_height) {
-			gs_stagesurface_destroy(src->stagesurface);
-			src->stagesurface = gs_stagesurface_create(width, height, GS_BGRA);
-			src->known_width = width;
-			src->known_height = height;
-		}
-
-		PROFILE_END(prof_render_target_name);
-
-		if (src->bypass_vectorscope) {
-			gs_blend_state_pop();
-
-			goto end;
-		}
-
-		gs_texrender_reset(src->texrender_uv);
-		if (vss_effect && gs_texrender_begin(src->texrender_uv, width, height)) {
-			PROFILE_START(prof_convert_yuv_name);
-			gs_ortho(0.0f, (float)width, 0.0f, (float)height, -100.0f, 100.0f);
-
-			gs_effect_t *effect = vss_effect;
-			gs_texture_t *tex = gs_texrender_get_texture(src->texrender);
-			if (tex) {
-				gs_effect_set_texture(gs_effect_get_param_by_name(effect, "image"), tex);
-				while (gs_effect_loop(effect, src->colorspace_calc==1 ? "ConvertRGB_UV601" : "ConvertRGB_UV709")) {
-					gs_draw_sprite(tex, 0, width, height);
-				}
-			}
-			gs_texrender_end(src->texrender_uv);
-			PROFILE_END(prof_convert_yuv_name);
-
-			PROFILE_START(prof_stage_surface_name);
-			gs_stage_texture(src->stagesurface, gs_texrender_get_texture(src->texrender_uv));
-			PROFILE_END(prof_stage_surface_name);
-			uint8_t *video_data = NULL;
-			uint32_t video_linesize;
-			PROFILE_START(prof_draw_vectorscope_name);
-			if (gs_stagesurface_map(src->stagesurface, &video_data, &video_linesize)) {
-				vss_draw_vectorscope(src, video_data, video_linesize);
-				gs_stagesurface_unmap(src->stagesurface);
-			}
-			PROFILE_END(prof_draw_vectorscope_name);
-		}
-		gs_blend_state_pop();
-	}
-	else
-		PROFILE_END(prof_render_target_name);
-
-end:
-	if (target)
-		obs_source_release(target);
 }
 
 static void create_graticule_vbuf(struct vss_source *src)
@@ -417,7 +237,7 @@ static void create_graticule_vbuf(struct vss_source *src)
 			{  63,  52 }, { 147,  44 }, {  44, 136 },
 		},
 	};
-	const int ppi = src->colorspace_calc-1;
+	const int ppi = src->cm.colorspace-1;
 
 	// label
 	for (int i=0; i<6; i++) {
@@ -457,7 +277,7 @@ static void create_graticule_vbuf(struct vss_source *src)
 	int stl_b = src->graticule_skintone_color >> 16 & 0xFF;
 	int stl_g = src->graticule_skintone_color >> 8 & 0xFF;
 	int stl_r = src->graticule_skintone_color & 0xFF;
-	switch(src->colorspace_calc) {
+	switch(src->cm.colorspace) {
 		case 1: // BT.601
 			stl_u = RGB2U_601(stl_r, stl_g, stl_b);
 			stl_v = RGB2V_601(stl_r, stl_g, stl_b);
@@ -493,21 +313,26 @@ static void vss_render(void *data, gs_effect_t *effect)
 {
 	UNUSED_PARAMETER(effect);
 	struct vss_source *src = data;
-	PROFILE_START(src->bypass_vectorscope ? prof_render_name_b : prof_render_name);
+	if (src->cm.bypass) {
+		cm_render_bypass(&src->cm);
+		return;
+	}
 
-	if (src->update_graticule || src->colorspace_calc<1) {
-		src->colorspace_calc = src->colorspace;
+	PROFILE_START(prof_render_name);
+
+	if (src->update_graticule || src->cm.colorspace<1) {
+		src->cm.colorspace = src->colorspace;
 		src->update_graticule = 0;
-		if (src->colorspace_calc<1 || 2<src->colorspace_calc) {
+		if (src->cm.colorspace<1 || 2<src->cm.colorspace) {
 			struct obs_video_info ovi;
 			if (obs_get_video_info(&ovi)) {
 				switch (ovi.colorspace) {
 					case VIDEO_CS_601:
-						src->colorspace_calc = 1;
+						src->cm.colorspace = 1;
 						break;
 					case VIDEO_CS_709:
 					default:
-						src->colorspace_calc = 2;
+						src->cm.colorspace = 2;
 						break;
 				}
 			}
@@ -518,23 +343,22 @@ static void vss_render(void *data, gs_effect_t *effect)
 		src->graticule_line_vbuf = NULL;
 	}
 
-	vss_render_target(src);
+	bool updated = cm_render_target(&src->cm);
 
-	if (src->bypass_vectorscope) {
-		gs_effect_t *effect = obs_get_base_effect(OBS_EFFECT_DEFAULT);
-		gs_texture_t *tex = gs_texrender_get_texture(src->texrender);
-		if (!tex)
-			goto end;
-		gs_effect_set_texture(gs_effect_get_param_by_name(effect, "image"), tex);
-		while (gs_effect_loop(effect, "Draw")) {
-			gs_draw_sprite(tex, 0, src->known_width, src->known_height);
+	if (updated) {
+		uint8_t *video_data = NULL;
+		uint32_t video_linesize;
+		PROFILE_START(prof_draw_vectorscope_name);
+		if (gs_stagesurface_map(src->cm.stagesurface, &video_data, &video_linesize)) {
+			vss_draw_vectorscope(src, video_data, video_linesize);
+			gs_stagesurface_unmap(src->cm.stagesurface);
 		}
-		goto end;
+		PROFILE_END(prof_draw_vectorscope_name);
 	}
 
 	PROFILE_START(prof_draw_name);
 	if (src->tex_vs) {
-		gs_effect_t *effect = vss_effect ? vss_effect : obs_get_base_effect(OBS_EFFECT_DEFAULT);
+		gs_effect_t *effect = cm_rgb2yuv_effect ? cm_rgb2yuv_effect : obs_get_base_effect(OBS_EFFECT_DEFAULT);
 		gs_effect_set_texture(gs_effect_get_param_by_name(effect, "image"), src->tex_vs);
 		gs_effect_set_float(gs_effect_get_param_by_name(effect, "intensity"), (float)src->intensity);
 		gs_effect_set_default(gs_effect_get_param_by_name(effect, "color"));
@@ -547,7 +371,7 @@ static void vss_render(void *data, gs_effect_t *effect)
 	PROFILE_START(prof_draw_graticule_name);
 	if (src->graticule_img.loaded && src->graticule) {
 		create_graticule_vbuf(src);
-		gs_effect_t *effect = vss_effect ? vss_effect : obs_get_base_effect(OBS_EFFECT_DEFAULT);
+		gs_effect_t *effect = cm_rgb2yuv_effect ? cm_rgb2yuv_effect : obs_get_base_effect(OBS_EFFECT_DEFAULT);
 		gs_effect_set_color(gs_effect_get_param_by_name(effect, "color"), src->graticule_color);
 		draw_uv_vbuffer(src->graticule_vbuf, src->graticule_img.texture, effect, "DrawGraticule", N_GRATICULES*2);
 	}
@@ -562,40 +386,7 @@ static void vss_render(void *data, gs_effect_t *effect)
 	}
 	PROFILE_END(prof_draw_graticule_name);
 
-end:;
-	PROFILE_END(src->bypass_vectorscope ? prof_render_name_b : prof_render_name);
-}
-
-static void vss_tick(void *data, float unused)
-{
-	UNUSED_PARAMETER(unused);
-	struct vss_source *src = data;
-
-	pthread_mutex_lock(&src->target_update_mutex);
-	if (src->target_name && !*src->target_name) {
-		if (src->weak_target)
-			obs_weak_source_release(src->weak_target);
-		src->weak_target = NULL;
-	}
-	if (is_preview_name(src->target_name)) {
-		obs_source_t *target = obs_frontend_get_current_preview_scene();
-		if (src->weak_target)
-			obs_weak_source_release(src->weak_target);
-		src->weak_target = target ? obs_source_get_weak_source(target) : NULL;
-		obs_source_release(target);
-	}
-	else if (src->target_name && *src->target_name && !src->weak_target && src->target_check_time) {
-		uint64_t t = os_gettime_ns();
-		if (t - src->target_check_time > SOURCE_CHECK_NS) {
-			src->target_check_time = t;
-			obs_source_t *target = obs_get_source_by_name(src->target_name);
-			src->weak_target = target ? obs_source_get_weak_source(target) : NULL;
-			obs_source_release(target);
-		}
-	}
-	pthread_mutex_unlock(&src->target_update_mutex);
-
-	src->rendered = 0;
+	PROFILE_END(prof_render_name);
 }
 
 struct obs_source_info colormonitor_vectorscope = {
@@ -610,7 +401,7 @@ struct obs_source_info colormonitor_vectorscope = {
 	.get_properties = vss_get_properties,
 	.get_width = vss_get_width,
 	.get_height = vss_get_height,
-	.enum_active_sources = vss_enum_sources,
+	.enum_active_sources = cm_enum_sources,
 	.video_render = vss_render,
-	.video_tick = vss_tick,
+	.video_tick = cm_tick,
 };
